requirements.txt
Django
matplotlib
numpy
gunicorn
dj-database-url
psycopg2
django-heroku
djangorestframework